+++ src/ocr/types.ts
@@ -6,10 +6,10 @@ export type IoE = Io & {
     readonly fetch: (url: string, options: RequestInit) => Promise<Response>
     readonly atob: (data: string) => string
     readonly log: (message: string) => void
-    readonly debug: (message: string, data?: any) => void
-    readonly warn: (message: string, data?: any) => void
-    readonly error: (message: string, error?: any) => void
-    readonly trace: (source: string, methodName: string, args?: any) => void
+    readonly debug: (message: string, data?: unknown) => void
+    readonly warn: (message: string, data?: unknown) => void
+    readonly error: (message: string, error?: unknown) => void
+    readonly trace: (source: string, methodName: string, args?: unknown) => void
 }
 
 /**